feat: view, viewModel 구현

+++ lib/data/mapper/upcoming_mapper.dart
@@ -20,7 +20,7 @@ extension UpcomingMapper on Results {
       title: title ?? '',
       overview: overview ?? '',
       releaseDate: releaseDate ?? '',
-      voteAverage: voteAverage?.toInt() ?? -1,
+      voteAverage: voteAverage?.toDouble() ?? -1,
     );
   }
 }
