+++ lib/data/model/now_playing.dart
@@ -14,7 +14,7 @@ class NowPlaying with _$NowPlaying {
     required String title,
     required String overview,
     required String releaseDate,
-    required int voteAverage,
+    required double voteAverage,
   }) = _NowPlaying;
 
   factory NowPlaying.fromJson(Map<String, Object?> json) => _$NowPlayingFromJson(json);
